chore: rename textCSS file
